Switch build system to microbundle
* Move interface/type definitinos to index.ts to not create unneeded files
* Fix example
