+++ examples/index.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
-  <script src="../dist/index.js"></script>
+  <script src="../lib/mesa-js-client.umd.js"></script>
   <script>
     const client = new MesaClient('ws://localhost:4500')
