--- tsconfig.browser.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "outDir": "./lib/browser",
-    "target": "es5",
-    "module": "es2015",
-    "moduleResolution": "node",
-    "lib": ["es2016", "dom", "es5"],
-    "allowSyntheticDefaultImports": true
-  },
-  "include": ["./src/client.ts"]
-}
